requirements.txt
streamlit==1.37.1
Pillow==10.4.0
pandas==2.2.2
requests==2.32.3
python-dateutil==2.9.0
numpy==1.26.4